travis: try qt-5.6 on OSX

+++ .travis-functions.sh
@@ -28,9 +28,11 @@ function script_generic
 {
 	xconfigure || return
 	$MAKE || return
-	./bin/DirectionalityIndicator &
-	sleep 10
-	kill $!
+	if test "$USE_QT" != "qt4"; then
+		./bin/DirectionalityIndicator &
+		sleep 20
+		kill $!
+	fi
 }
 
 function install_deps_linux
@@ -61,8 +63,10 @@ function install_deps_osx
 	brew update >/dev/null
 	brew install \
 		glew \
-		qt5 \
 		|| return
+
+	curl -O https://raw.githubusercontent.com/Homebrew/homebrew-core/fdfc724dd532345f5c6cdf47dc43e99654e6a5fd/Formula/qt5.rb
+	brew install ./qt5.rb
 
 	QT_DIR=$(brew --prefix qt5)
 	PATH="$QT_DIR/bin:$PATH"
